log: clean up / extend the color code
This patch extends output colorization to more terminal types than
those previously supported.

I also took the opportunity to extract color-related code to a
separate file.
